refactor: Change daily view toolbar link location
Changes the link location of the daily view side bar links to
redirect to the corresponding month view since this feels more
intuitive. There's fewer cases where users want to jump to the
start of the month in the daily view and all of these can be
addressed by going to the month view before selecting the date there.