style: Trim height of habit tracker page
Trims height of habit tracker page to fit display and to match the
other pages. This was an oversight when shrinking the grid boxes
to provide more room for description labels. Although this
preserved the width, the height increased slightly. The difference
in height was removed from the grid header boxes to compensate.